third_party/WebKit/LayoutTests/webaudio/audiobuffer-getChannelData-expected.txt

Test AudioBuffer.getChannelData() Returns the Same Object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS a = buffer.getChannelData(0)
PASS b = buffer.getChannelData(0)
PASS a === b is true
PASS a = buffer.getChannelData(1)
PASS b = buffer.getChannelData(1)
PASS a === b is true
PASS getChannelData correctly returned the same buffer.

PASS a = buffer1.getChannelData(0)
PASS b = buffer2.getChannelData(0)
PASS a === b is false
PASS a = buffer1.getChannelData(1)
PASS b = buffer2.getChannelData(1)
PASS a === b is false
PASS getChannelData correctly returned different buffers.

PASS successfullyParsed is true

TEST COMPLETE
